+++ source/observers/mutation.js
@@ -4,6 +4,7 @@
  */
 
 import presets from '../lib/presets';
+import * as observers from '../configs/observers';
 
 const DOM_NODE_INSERTED = 'DOMNodeInserted';
 const DOM_NODE_REMOVED = 'DOMNodeRemoved';
@@ -13,40 +14,44 @@ const DOM_SUBTREE_MODIFIED = 'DOMSubtreeModified';
 export default (container) => {
     let attached;
     let mutationObserver;
+    let _callback;
 
     const on = (callback) => {
         if (attached) {
             return false;
         } else {
             attached = true;
+            _callback = () => {
+                return callback(observers.MUTATION);
+            };
             presets(container);
             const MutationObserver = window.MutationObserver || window.WebKitMutationObserver;
             if (MutationObserver) {
-                mutationObserver = new MutationObserver(callback);
+                mutationObserver = new MutationObserver(_callback);
                 mutationObserver.observe(container, {
                     childList: true,
                     attributes: true,
                     subtree: true,
                 });
             } else {
-                container.addEventListener(DOM_NODE_INSERTED, callback);
-                container.addEventListener(DOM_NODE_REMOVED, callback);
-                container.addEventListener(DOM_ATTR_MODIFIED, callback);
-                container.addEventListener(DOM_SUBTREE_MODIFIED, callback);
+                container.addEventListener(DOM_NODE_INSERTED, _callback);
+                container.addEventListener(DOM_NODE_REMOVED, _callback);
+                container.addEventListener(DOM_ATTR_MODIFIED, _callback);
+                container.addEventListener(DOM_SUBTREE_MODIFIED, _callback);
             }
             return true;
         }
     };
 
-    const off = (callback) => {
+    const off = () => {
         if (attached) {
             if (mutationObserver) {
                 mutationObserver.disconnect();
             } else {
-                container.removeEventListener(DOM_NODE_INSERTED, callback);
-                container.removeEventListener(DOM_NODE_REMOVED, callback);
-                container.removeEventListener(DOM_ATTR_MODIFIED, callback);
-                container.removeEventListener(DOM_SUBTREE_MODIFIED, callback);
+                container.removeEventListener(DOM_NODE_INSERTED, _callback);
+                container.removeEventListener(DOM_NODE_REMOVED, _callback);
+                container.removeEventListener(DOM_ATTR_MODIFIED, _callback);
+                container.removeEventListener(DOM_SUBTREE_MODIFIED, _callback);
             }
             attached = false;
             return true;