+++ CHANGELOG.md
@@ -1,5 +1,11 @@
-# 1.0.1
+# 1.1.0 more APIs, more accuracy when telling impression changed
 
-- fix event callback data
+- added types into callback arguments
+- added observers events
+- `end` event now triggered by `beforeunload` browser event
+
+# 1.0.1 bugfix
+
+- fixed event callback data
 
 # 1.0.0 first version